+++ pyroomacoustics/datasets/__init__.py
@@ -141,3 +141,4 @@ from .base import Meta, Sample, AudioSample, Dataset
 from .timit import Word, Sentence, TimitCorpus
 from .cmu_arctic import CMUArcticCorpus, CMUArcticSentence, cmu_arctic_speakers
 from .google_speech_commands import GoogleSpeechCommands, GoogleSample
+from .locata import LOCATA, LocataRecording